# libct/cg/sd/v1: Fix unnecessary freeze/thaw