My testing shows that https://github.com/opencontainers/runc/pull/3082 (commit https://github.com/opencontainers/runc/pull/3082/commits/f2db87986cab5a7c98b8d9b875bb54e0e7a44219) doesn't work as expected. This causes big issues in k8s.

PTAL @kolyshkin :smile: 

Finally back again after having some time off, so haven't had time to look at it before now.