@kolyshkin 2021-08-10
I guess what we need first is a unit test that would create a unit (of various types) with a particular property, get the property, and check it's as expected.

Last commit of #3082 is an integration test (`TestPodSkipDevicesUpdate`) that checks that the freeze is not done. I look at all this and I still can't understand why that test does not catch the issue this PR is fixing...