@odinuge 2021-08-11
>I guess what we need first is a unit test that would create a unit (of various types) with a particular property, get the property, and check it's as expected.

Yeah, some tests would be ok I guess. It is fairly hard to test tho. But keep reading `freezer.state` on the slice while constantly updating, and looking for only THAWED would work. I can look into doing it in go, but the bats stuff is a bit complicated tho., so that would require some boiler plate.

> Last commit of #3082 is an integration test (TestPodSkipDevicesUpdate) that checks that the freeze is not done. I look at all this and I still can't understand why that test does not catch the issue this PR is fixing...

Well, it only verifies that there is no errors with the devices. It does however not check if the slice is frozen or not.


To test this, one can simply verify like this - keep the loop running while running tests with `sudo go test -run=TestPodSkipDevicesUpdate -v -test.count=5`:

```
# in this branch
$ while true ; do grep FROZEN -R /sys/fs/cgroup/freezer/system.slice/; done
/sys/fs/cgroup/freezer/system.slice/system-runc_test_podTestPodSkipDevicesUpdate.slice/test-PodSkipDevicesUpdate.scope/freezer.state:FROZEN
/sys/fs/cgroup/freezer/system.slice/system-runc_test_podTestPodSkipDevicesUpdate.slice/test-PodSkipDevicesUpdate.scope/freezer.state:FROZEN
# in master 
$ while true ; do grep FROZEN -R /sys/fs/cgroup/freezer/system.slice/; done             
/sys/fs/cgroup/freezer/system.slice/system-runc_test_podTestPodSkipDevicesUpdate.slice/test-PodSkipDevicesUpdate.scope/freezer.state:FROZEN                     
/sys/fs/cgroup/freezer/system.slice/system-runc_test_podTestPodSkipDevicesUpdate.slice/freezer.state:FROZEN                                                     
/sys/fs/cgroup/freezer/system.slice/system-runc_test_podTestPodSkipDevicesUpdate.slice/test-PodSkipDevicesUpdate.scope/freezer.state:FROZEN                     
/sys/fs/cgroup/freezer/system.slice/system-runc_test_podTestPodSkipDevicesUpdate.slice/freezer.state:FROZEN                                                     
/sys/fs/cgroup/freezer/system.slice/system-runc_test_podTestPodSkipDevicesUpdate.slice/test-PodSkipDevicesUpdate.scope/freezer.state:FROZEN                     
/sys/fs/cgroup/freezer/system.slice/system-runc_test_podTestPodSkipDevicesUpdate.slice/freezer.state:FROZEN                                                     
/sys/fs/cgroup/freezer/system.slice/system-runc_test_podTestPodSkipDevicesUpdate.slice/test-PodSkipDevicesUpdate.scope/freezer.state:FROZEN                     
/sys/fs/cgroup/freezer/system.slice/system-runc_test_podTestPodSkipDevicesUpdate.slice/freezer.state:FROZEN                                                     
/sys/fs/cgroup/freezer/system.slice/system-runc_test_podTestPodSkipDevicesUpdate.slice/test-PodSkipDevicesUpdate.scope/freezer.state:FROZEN                     
/sys/fs/cgroup/freezer/system.slice/system-runc_test_podTestPodSkipDevicesUpdate.slice/freezer.state:FROZEN                                                     
/sys/fs/cgroup/freezer/system.slice/system-runc_test_podTestPodSkipDevicesUpdate.slice/test-PodSkipDevicesUpdate.scope/freezer.state:FROZEN                     
/sys/fs/cgroup/freezer/system.slice/system-runc_test_podTestPodSkipDevicesUpdate.slice/freezer.state:FROZEN                                                     
/sys/fs/cgroup/freezer/system.slice/system-runc_test_podTestPodSkipDevicesUpdate.slice/test-PodSkipDevicesUpdate.scope/freezer.state:FROZEN                     
/sys/fs/cgroup/freezer/system.slice/system-runc_test_podTestPodSkipDevicesUpdate.slice/freezer.state:FROZEN                                                     
/sys/fs/cgroup/freezer/system.slice/system-runc_test_podTestPodSkipDevicesUpdate.slice/test-PodSkipDevicesUpdate.scope/freezer.state:FROZEN                     
/sys/fs/cgroup/freezer/system.slice/system-runc_test_podTestPodSkipDevicesUpdate.slice/freezer.state:FROZEN                                                     
/sys/fs/cgroup/freezer/system.slice/system-runc_test_podTestPodSkipDevicesUpdate.slice/freezer.state:FROZEN                        
```

As seen here, only the scope/container is frozen with this branch, while in master both the pod/slice and the container/scope is frozen. Does that make sense to you @kolyshkin ?