@kolyshkin 2021-08-11
> Hmm. Is something wrong with CI, or?

Yup; see https://githubstatus.com/.

Seems over now, do a force push to trigger CI.