@odinuge 2021-08-11
> @AkihiroSuda: Does this fix kubernetes/kubernetes#104280 ?

Yes.

But as I have said a few times already, for k8s a flag for disabling this would be better and simpler. I guess we (in k8s, not here in runc) have to discuss if we should avoid using `runc/libcontainer` for setting cgroup things, since things like this will increase the overhead waaay more than necessary.