@kolyshkin 2021-08-11
@odinuge I was playing with the test case and ended up doing https://github.com/opencontainers/runc/pull/3148.

The other thing I did is slightly simplified the deviceAllow check (https://github.com/opencontainers/runc/pull/3143#discussion_r687222469)

We can either use #3148, or cherry-pick the test case from it to here.