@odinuge 2021-08-11
> We can either use #3148, or cherry-pick the test case from it to here.

I can cherry-pick the test commit here if they should be in the same PR. I prefer to keep this PR, otherwise I cannot update and change the PR (you created based of this), making it a bit cumbersome to work on it. Having a separate PR with the test works as well.