@kolyshkin 2021-08-11
> I can cherry-pick the test commit here if they should be in the same PR

Please do (otherwise the CI is semi-useless).

> making it a bit cumbersome to work on it

Something like `git-rebase -i origin/master` is a life saver when working on PRs with multiple commits. In most complicated cases I do `git stash; git format-patch -o <dir> origin/master; git reset --hard origin/master`, when `git am` those commits one by one, making changes and doing  `git commit --amend` when necessary.