@kolyshkin 2021-08-11
Once ready and merged, we'll have a backport to 1.0 and will try 1.0.2 released ASAP.