@kolyshkin 2021-08-12
> things like this will increase the overhead waaay more than necessary

@odinuge your concern about this code path being slow is valid, it is also quite complicated. I was looking into how to make it fast and simple -- PTAL https://github.com/opencontainers/runc/pull/3151.

Surely a simple "SkipFreeze" flag will be faster, but this seems lightweight enough to not cause performance concerns.