@kolyshkin 2021-08-13
OK I ended up writing this benchmark for `freezeBeforeSet()` (based on the test code):

Here are the results (first line is this PR, second is #3151):
```
BenchmarkFreezeBeforeSet-8   	    3757	    286971 ns/op	   16632 B/op	     680 allocs/op
BenchmarkFreezeBeforeSet-8   	  131257	      9116 ns/op	    2344 B/op	       7 allocs/op
```

To have some reference points, I did a few quick benchmarks

1. start/stop a systemd scope:
```
BenchmarkApplyDestroy-8      	     808	   1390820 ns/op	  172188 B/op	    2130 allocs/op
``` 

2. Set() (with no freeze checking or freeze code -- as if SkipFreeze flag is provided)
```
BenchmarkSet-8               	    6412	    183464 ns/op	    7266 B/op	     283 allocs/op
```

3. Set() (with unconditional freeze, no processes in cgroup!):
```
BenchmarkSet-8               	    5127	    231167 ns/op	   10224 B/op	     298 allocs/op
```

From that we can estimate Set() overhead for kubelet case

* with current approach (#3082 (commit f2db879) + #3143): slowdown from 200 to 500 us (150% worse)
* with devices.list approach (#3151): slowdown from 200 to 210 us (5% worse)